refactor: 예약 관리 1-3. 패키지 구조 변경
- controller, dto, entity, repository, service로 패키지 구분
