+++ src/main/java/roomescape/admin/dto/ReadReservationTimeResponse.java
@@ -1,4 +1,6 @@
-package roomescape.admin;
+package roomescape.admin.dto;
+
+import roomescape.admin.entity.ReservationTime;
 
 import java.util.List;
 import java.util.stream.Collectors;
