+++ src/main/java/roomescape/admin/dto/SaveReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.admin;
+package roomescape.admin.dto;
 
 public record SaveReservationTimeRequest(String startAt) {
 }
