+++ src/main/java/roomescape/admin/entity/Reservation.java
@@ -1,4 +1,6 @@
-package roomescape.admin;
+package roomescape.admin.entity;
+
+import roomescape.admin.dto.SaveReservationRequest;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
