+++ src/main/java/roomescape/admin/repository/ReservationTimeQuery.java
@@ -1,4 +1,4 @@
-package roomescape.admin;
+package roomescape.admin.repository;
 
 public class ReservationTimeQuery {
     public static final String READ_ALL = "select id as time_id, start_at as time_start_at from reservation_time";
